backend/requirements.txt
# NeuroFinance Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# NLP & ML
transformers==4.37.0
sentencepiece==0.1.99

# Data Processing
feedparser==6.0.10
python-dateutil==2.8.2

# HTTP & WebSockets
aiohttp
websockets
httpx

# Utilities
python-multipart
python-dotenv

# Optional: Monitoring & Logging
prometheus-client==0.19.0
